# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NPI Website API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme."
      }
    },
...
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate user and get access token",
        "description": "Sends credentials to Tesla Factory API to validate user and return a JWT token. The token is required for subsequent API calls. Rate limited to 5 attempts per minute.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@tesla.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Successfully authenticated and redirected to home page",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "session=encrypted_session_data; HttpOnly; Secure; SameSite=Lax; Path=/"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid username or password"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/factory/ar/mass-hold": {
      "post": {
        "summary": "Put multiple things on hold in AR containment",
        "description": "Places multiple things on hold in a specified AR containment by importing them as suspects",
        "tags": ["Factory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["thingNames", "containmentName", "token"],
                "properties": {
                  "thingNames": {
                    "type": "string",
                    "description": "Comma-separated list of thing names to put on hold",
                    "example": "THING001,THING002,THING003"
                  },
                  "containmentName": {
                    "type": "string",
                    "description": "Name of the AR containment",
                    "example": "AR_CONTAINMENT_001"
                  },
                  "token": {
                    "type": "string",
                    "description": "JWT authentication token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Things put on hold successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Thing(s) put on hold successfully"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "successful": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "List of successfully processed thing names"
                        },
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "thingName": {
                                "type": "string"
                              },
                              "error": {
                                "type": "string"
                              }
                            }
                          },
                          "description": "List of failed operations with error details"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/flow/get-flow/{slug}": {
      "get": {
        "summary": "Get production flows for a process",
        "description": "Retrieves all production flows for a specified process name",
        "tags": ["Factory"],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Process name to get flows for",
            "example": "PROCESS_001"
          }
        ],
        "responses": {
          "200": {
            "description": "Production flows retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully fetched flows"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "flows": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          },
                          "description": "Array of production flows"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}